requirements.txt
flask==2.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.1
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
jinja2==3.0.3
flask-restx==0.5.1
Werkzeug==2.0.3
python-dotenv==0.19.2
Flask-Minify==0.37
Flask-Dance==5.1.0
blinker==1.4
pyOpenSSL==23.2.0
PyJWT==2.4.0
marshmallow==3.15.0
flask-cors==3.0.10
# flask_mysqldb
